--- packages/@aws-c2a/broadening-permissions/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './security-changes';
