+++ packages/@aws-c2a/presets/babel.config.js
