+++ packages/@aws-c2a/presets/fixtures/broadening-iam-permissions.json
