+++ packages/@aws-c2a/presets/fixtures/broadening-permissions.json
