+++ packages/@aws-c2a/presets/fixtures/broadening-security-group.json
