+++ packages/@aws-c2a/presets/fixtures/create-fixtures.ts
@@ -1,10 +1,10 @@
 import { writeFileSync } from 'fs';
 import { resolve } from 'path';
-import { SecurityChangesRules } from '../lib';
+import { IamChanges, SecurityGroup, PresetRules } from '../lib';
 
-const iamRules = SecurityChangesRules.BroadeningIamPermissions();
-const securityGroupRules = SecurityChangesRules.BroadeningSecurityGroup();
-const broadeningPermissions = new SecurityChangesRules(...iamRules.rules, ...securityGroupRules.rules);
+const iamRules = IamChanges.BroadeningPermissions();
+const securityGroupRules = SecurityGroup.BroadeningPermissions();
+const broadeningPermissions = new PresetRules(...iamRules.rules, ...securityGroupRules.rules);
 
 writeFileSync(resolve(__dirname, 'broadening-iam-permissions.json'), iamRules.toString());
 writeFileSync(resolve(__dirname, 'broadening-security-group.json'), securityGroupRules.toString());
