+++ packages/@aws-c2a/presets/lib/generators/path.ts
