+++ packages/@aws-c2a/presets/lib/iam-changes.ts
@@ -1,6 +1,7 @@
 import { RuleRisk } from '@aws-c2a/models';
-import { CUserRules, Change, ChangeType, Rule } from '@aws-c2a/rules';
+import { Change, ChangeType, Rule } from '@aws-c2a/rules';
 import { generateComponent, generateHighRiskChild, generateStatementRules } from './generators';
+import { PresetRules } from './preset-rules';
 import {
   IAM_INLINE_IDENTITY_POLICIES,
   IAM_INLINE_RESOURCE_POLICIES,
@@ -9,40 +10,17 @@ import {
   IAM_POLICY_RESOURCES,
 } from './private/security-policies';
 
-export class SecurityChangesRules {
-  /**
-   * Rules that pertain to broadening permissions for
-   * EC2 security group changes.
-   */
-  public static BroadeningSecurityGroup(): SecurityChangesRules {
-    const rules = new SecurityChangesRules();
-    const rootBindings = new Rule([Change.ALL, Change.INSERT]);
-    const { component: securityGroup, rule: sgBindings } = generateComponent('AWS::EC2::SecurityGroup', rootBindings);
-    ['Ingress', 'Egress'].forEach(type => {
-      const { component, rule: componentBindings } = generateComponent(`AWS::EC2::SecurityGroup${type}`, rootBindings);
-
-      generateHighRiskChild(componentBindings, component, {
-        change: Change.ALL,
-      });
-      generateHighRiskChild(sgBindings, securityGroup, {
-        change: Change.INSERT,
-        equals: '*',
-        sourcePath: ['Properties', `SecurityGroup${type}`, '*'],
-      });
-      generateHighRiskChild(sgBindings, securityGroup, {
-        change: Change.ALL,
-        targetPath: ['Properties', `SecurityGroup${type}`, '*'],
-      });
-    });
-    rules.addRules(rootBindings.toJSON());
-    return rules;
-  }
+/**
+ * A class containing static presets that represent levels
+ * of permissions.
+ */
+export class IamChanges extends PresetRules {
   /**
    * Rules that pertain to broadening permissions for
    * IAM policy/statement changes.
    */
-  public static BroadeningIamPermissions(): SecurityChangesRules {
-    const rules = new SecurityChangesRules();
+  public static BroadeningPermissions(): PresetRules {
+    const rules = new IamChanges();
     const rootBindings = new Rule([Change.INSERT, Change.INSERT_PROP, Change.UPDATE_PROP]);
 
     // Lambda Permissions
@@ -107,33 +85,6 @@ export class SecurityChangesRules {
 
     rules.addRules(rootBindings.toJSON());
     return rules;
-  }
-  public static BroadeningPermissions(): SecurityChangesRules {
-    const rules = new SecurityChangesRules();
-    rules.addRules(
-      ...SecurityChangesRules.BroadeningSecurityGroup().rules,
-      ...SecurityChangesRules.BroadeningIamPermissions().rules,
-    );
-    return rules;
-  }
-
-  private _rules: CUserRules = [];
-  get rules(): CUserRules {
-    return this._rules;
-  }
-
-  constructor(...rules: CUserRules) {
-    this.addRules(...rules);
-  }
-
-  public addRules(...rules: CUserRules): void {
-    this._rules.push(...rules);
-  }
-
-  public toString(prettier = false): string {
-    return prettier
-      ? JSON.stringify(this.rules, null, 2)
-      : JSON.stringify(this.rules);
   }
 }
