+++ packages/@aws-c2a/presets/lib/index.ts
@@ -1,0 +1,3 @@
+export * from './iam-changes';
+export * from './preset-rules';
+export * from './security-group';
