+++ packages/@aws-c2a/presets/lib/private/object.ts
