+++ packages/aws-c2a/lib/toolkit.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
-import { SecurityChangesRules } from '@aws-c2a/broadening-permissions';
 import { createChangeAnalysisReport, CDKParser } from '@aws-c2a/engine';
 import { ChangeAnalysisReport, groupArrayBy, JSONSerializer, RuleRisk, Transition } from '@aws-c2a/models';
+import { IamChanges, SecurityGroup } from '@aws-c2a/presets';
 import { CUserRules } from '@aws-c2a/rules';
 import { IC2AHost } from './c2a-host';
 import { CfnTraverser } from './cfn-traverser';
@@ -79,7 +79,9 @@ export class C2AToolkit {
     const before: {[stackName: string]: TemplateTree} = {};
     const after: {[stackName: string]: TemplateTree} = {};
     const rules = [
-      ...(options.broadeningPermissions ? SecurityChangesRules.BroadeningPermissions().rules : []),
+      ...(options.broadeningPermissions
+        ? [...IamChanges.BroadeningPermissions().rules, ...SecurityGroup.BroadeningPermissions().rules]
+        : []),
       ...(options.rulesPath ? JSON.parse(await fs.promises.readFile(options.rulesPath, 'utf-8')) : []),
     ];
