+++ packages/aws-c2a/package.json
@@ -30,7 +30,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-c2a/broadening-permissions": "^0.4.0",
+    "@aws-c2a/presets": "^0.4.0",
     "@aws-c2a/engine": "^0.4.0",
     "@aws-c2a/models": "^0.4.0",
     "@aws-c2a/rules": "^0.4.0",
